requirements.txt
absl-py==0.9.0
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.9.12
anaconda-project==0.8.2
asn1crypto==0.24.0
astroid==1.6.1
astropy==2.0.3
astunparse==1.6.3
attrs==17.4.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
bokeh==0.12.13
boto==2.48.0
boto3==1.14.11
botocore==1.17.11
Bottleneck==1.2.1
cachetools==4.1.0
certifi==2020.6.20
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda==4.8.3
conda-build==3.4.1
conda-package-handling==1.7.0
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.9.2
cvxopt==1.2.7
cvxpy==1.1.18
cycler==0.10.0
Cython==0.29.14
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.4.2
distributed==1.20.2
docutils==0.14
drawnow==0.72.0
ecos==2.0.10
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.13
Flask==0.12.2
Flask-Cors==3.0.3
gast==0.3.3
gensim==3.8.3
gevent==1.2.2
glob2==0.6
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.12
grpcio==1.30.0
h5py==2.10.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
importlib-metadata==1.6.1
ipython==6.2.1
ipython_genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jmespath==0.10.0
joblib==0.15.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.4
jupyterlab-launcher==0.10.2
kaggle @ file:///C:/Users/Administrator/Desktop/kaggle-1.5.6.tar.gz
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.3.1
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.11
mistune==0.8.3
mpmath==1.0.0
msgpack-python==0.5.1
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.4
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpy==1.19.5
numpydoc==0.7.0
oauthlib==3.1.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.4.10
opt-einsum==3.2.1
osqp==0.6.2.post5
packaging==16.8
pandas==1.0.5
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.12.2
psutil==5.4.3
py==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.5
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.3.2
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2017.3
PyWavelets==0.5.2
pywin32==222
pywinpty==0.5
PyYAML==5.3.1
pyzmq==16.0.3
qdldl==0.1.5.post0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
requests==2.21.0
requests-oauthlib==1.3.0
rope==0.10.7
rsa==4.6
ruamel_yaml==0.15.35
s3transfer==0.3.3
scikit-image==0.13.1
scikit-learn==0.22.2.post1
scikit-survival==0.12.0
scipy==1.5.0
scs==3.1.0
seaborn==0.8.1
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
smart-open==2.0.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.3
threadpoolctl==2.1.0
toolz==0.9.0
tornado==4.5.3
tqdm==4.46.1
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
win-unicode-console==0.5
win_inet_pton==1.0.1
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.3.0
xmltodict==0.12.0
zict==0.1.3
zipp==3.1.0
